change type of instruction count(~

+++ base.c
@@ -19,7 +19,7 @@ unsigned int codesize;
 int stop = 0;//コードの実行を中止/終了
 int mode_jump = 0;//次の命令までジャンプするモード
 int mode_sipnext = -1;//次の番地の命令まで続けて実行するモード。SIP時に使う
-int used[NUM_OF_OP];//各命令実行回数
+long used[NUM_OF_OP];//各命令実行回数
 unsigned long dyna = 0;//動的実行命令数
 int branch[NUM_OF_OP];
 int nbranch[NUM_OF_OP];
